Checking and tweaks to allowable input values
Adjusted allowable input values so that mannings could not be 0.0 and to allow smaller values for channel width and depth.  Also increased max bed slope to 1 in 1.  Bottom width can be less than top width, in real world this is a bug but the eqns don't care so not addressed this.  Tested the existing online tool against ISIS1D and generated identical flows for a range of sections.